requirements.txt
mysql-connector-python>=8.0.17
python-slugify>=3.0.6